added projected pattern to the chart
