+++ client/src/components/icons/projected-data.tsx
@@ -6,6 +6,7 @@ function ProjectedDataIcon(props) {
       viewBox="0 0 12 12"
       fill="none"
       xmlns="http://www.w3.org/2000/svg"
+      id="projected-data-icon"
       {...props}
     >
       <mask
