--- client/src/containers/analysis-chart/comparison-chart/tooltip/index.ts
@@ -1,1 +1,0 @@
-export { default } from './component';
