--- client/src/containers/analysis-chart/impact-chart/tooltip/index.ts
@@ -1,1 +1,0 @@
-export { default } from './component';
